buck2_util

Module cycle_detector

Structs

LazyCycleDetector

The LazyCycleDetector is used to detect cycles in a parallel graph traversal. In essence, it operates on a sequence of events for nodes that inform it when a node is “started”, an edge is discovered and a node is “finished” (which indicates that we won’t see any new edges for that node and that the entire graph below that node is guaranteed to have no cycles). After receiving any events, the cycle detector will wait for an idle period (without new events) and then traverse the active graph to check for cycles. Any nodes that are in or depend on a cycle will have their active LazyCycleDetectorGuard notified of the cycle. The guard provides a function guard_this that should be used to wrap futures where one “node” is waiting on dependencies, this will then return a future that will exit early with an error if a cycle is detected.

LazyCycleDetectorGuard

Traits

CycleDescriptor